Revert: "Core/Scripts: Completed transmogrify code." .. Pull Request coming.
This reverts commit 696513ee4100d029da7abe1d0e3cfc140cc53ea0.

+++ src/server/scripts/Custom/npc_transmogrify.cpp
@@ -11,7 +11,7 @@ enum TransmogrifyActions {
     ACTION_TRANSMOGRIFY_REMOVE_DISPLAY
 };
 
-const uint64 PriceInGold = 1000 * 100 * 100; // 1k golds
+const uint16 PriceInGold = 1000 * 100 * 100; // 1k golds
 
 class npc_transmogrify : public CreatureScript
 {
